Enable more expressive emails with `send_email_with_attachment`
This commit adds the ability to write more expressive e-mails
with the `send_email_with_attachment` function by enabling
direct `MIMEText` to be passed as the `body` of the e-mail to be
sent.

This allows for `MIMEText` with subtype `html` (or other subtypes)
which enables feature like messages with hyperlinks, etc.

Previously, the function only allowed passing `str` as the body
text which disable use of hyperlinks, etc.

See the following for examples of fancy e-mail features:
https://realpython.com/python-send-email/#sending-fancy-emails